chore: Remove folder for tx layouts. Move files one folder up
