--- source/components/Layout/Transaction/index.tsx
@@ -1,2 +1,0 @@
-export { TxLayout } from './TxLayout';
-export { TxConfirmLayout } from './TxConfirmLayout';
